Update YOLO process code
* Remove library layer for overlay node
* Simplify logic of bounding box Class
* Rename from yolo_processor to yolo_process
* Apply clang-format
