+++ cv_tensor_process/yolo_v8_process/qrb_ros_yolo_process/launch/__pycache__/component.launch.cpython-310.pyc
