+++ cv_tensor_process/yolo_v8_process/qrb_yolo_process_lib/package.xml
@@ -1,0 +1,12 @@
+<package format="2">
+  <name>qrb_yolo_process</name>
+  <version>0.0.1</version>
+  <description>Library for Yolo pre/post process</description>
+  <maintainer email="quic_xiaolee@quicinc.com">Xiao Li</maintainer>
+  <license>BSD-3-Clause-Clear</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
