--- cv_tensor_process/yolov8_object_detection_process/qrb_yolo_processor_lib/include/yolo_det_overlay.hpp
@@ -1,31 +1,0 @@
-// Copyright (c) 2024 Qualcomm Innovation Center, Inc. All rights reserved.
-// SPDX-License-Identifier: BSD-3-Clause-Clear
-
-#ifndef _QRB_YOLO_DETECTION_OVERLAY_HPP_
-#define _QRB_YOLO_DETECTION_OVERLAY_HPP_
-
-#include <vector>
-
-#include "common.hpp"
-#include "opencv2/core.hpp"
-#include "opencv2/imgproc.hpp"
-#include "opencv2/opencv.hpp"
-
-namespace qrb::yolo_processor
-{
-class YoloDetOverlay
-{
-public:
-  YoloDetOverlay() = default;
-  ~YoloDetOverlay() = default;
-
-  /**
-   * \brief draw bounding boxes on the given cv::Mat object in an in-place manner.
-   * \param instances Contains the detected instances info.
-   * \param img The instance info will be overlaid onto this image.
-   */
-  void draw_inplace(std::vector<YoloInstance> & instances, cv::Mat & img);
-};
-
-}  // namespace qrb::yolo_processor
-#endif  // _QRB_YOLO_DETECTION_OVERLAY_HPP_
