--- cv_tensor_process/yolov8_object_detection_process/qrb_yolo_processor_lib/package.xml
@@ -1,12 +1,0 @@
-<package format="2">
-  <name>qrb_yolo_processor</name>
-  <version>1.1.0</version>
-  <description>Library for Yolo pre/post process</description>
-  <maintainer email="quic_xiaolee@quicinc.com">Xiao Li</maintainer>
-  <license>BSD-3-Clause-Clear</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
